Fix failure when some jobs have only skipped lemmas
